Fixes for protos system.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1059 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: cc5c7b860a9164e8f1c86ec8e96d28e9c6f9b5b7

+++ Source/Client.cpp
@@ -3123,7 +3123,7 @@ void FOClient::Net_OnAddCritter( bool is_npc )
     Bin >> hy;
     Bin >> dir;
 
-    uchar cond;
+    int   cond;
     uint  anim1life, anim1ko, anim1dead;
     uint  anim2life, anim2ko, anim2dead;
     uint  flags;
